api/requirements.txt
google-cloud-aiplatform==1.49.0
